Revert to one Arrow example
